Implement email confirmation on user signup
Updated the user sign up flow to include email confirmation. The system now sends an activation link to the user's email after registration, and the account is only activated after the link is clicked. Added necessary URLs, views, and templates to support this flow. Also, new django allauth backend is added to support this authentication method.